add two packed printers
